warning pass, fix many -Wall, -Wextra warnings
 - many signed vs unsigned comparisons
 - some mixed signed/unsigned result ternary expressions
 - (void) cast some unused parameters
 - remove some unused parameters
 - add missing initializers to some structure initializations
 - restructure fill object initialization to only brace initialize
   the base structure
 - add /* FALLTHROUGH */ comments to a few switch statements to
   suppress some switch case fall through warnings

+++ hlines.c
@@ -184,7 +184,7 @@ i_int_hlines_add(i_int_hlines *hlines, i_img_dim y, i_img_dim x, i_img_dim width
     else {
       i_int_hline_seg *seg;
       /* add a new segment */
-      if (entry->count == entry->alloc) {
+      if ((size_t)entry->count == entry->alloc) {
 	/* expand it */
 	size_t alloc = entry->alloc * 3 / 2;
 	entry = myrealloc(entry, sizeof(i_int_hline_entry) +
